+++ packages/manager/src/components/HighlightedMarkdown/HighlightedMarkdown.tsx
@@ -1,18 +1,20 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-import { Theme, useTheme } from '@mui/material/styles';
 import * as hljs from 'highlight.js/lib/core';
 import apache from 'highlight.js/lib/languages/apache';
 import bash from 'highlight.js/lib/languages/bash';
 import javascript from 'highlight.js/lib/languages/javascript';
 import nginx from 'highlight.js/lib/languages/nginx';
 import yaml from 'highlight.js/lib/languages/yaml';
+import HLJSDarkTheme from 'highlight.js/styles/a11y-dark.css?raw';
+import HLJSLightTheme from 'highlight.js/styles/a11y-light.css?raw';
 import * as React from 'react';
 import sanitize from 'sanitize-html';
 
 import { Typography } from 'src/components/Typography';
 import 'src/formatted-text.css';
+import { ThemeName } from 'src/foundations/themes';
 import { unsafe_MarkdownIt } from 'src/utilities/markdown';
 import { sanitizeHTML } from 'src/utilities/sanitize-html';
+import { useColorMode } from 'src/utilities/theme';
 
 hljs.registerLanguage('apache', apache);
 hljs.registerLanguage('bash', bash);
@@ -37,9 +39,38 @@ export interface HighlightedMarkdownProps {
 }
 
 export const HighlightedMarkdown = (props: HighlightedMarkdownProps) => {
-  const theme = useTheme<Theme>();
   const { className, language, sanitizeOptions, textOrMarkdown } = props;
   const rootRef = React.useRef<HTMLDivElement>(null);
+
+  const { colorMode } = useColorMode();
+
+  /**
+   * This function exists because we use Hightlight.js and it does not have a built-in
+   * way to programaticly change the theme.
+   *
+   * We must manually switch our Highlight.js theme's CSS when our theme is changed.
+   */
+  const handleThemeChange = async (theme: ThemeName) => {
+    const THEME_STYLE_ID = 'hljs-theme';
+    const existingStyleTag = document.getElementById(THEME_STYLE_ID);
+
+    if (existingStyleTag) {
+      // If the style tag already exists in the <head>, just update the css content.
+      existingStyleTag.innerHTML =
+        theme === 'light' ? HLJSLightTheme : HLJSDarkTheme;
+    } else {
+      // The page has been loaded and we need to manually append our Hightlight.js
+      // css so we can easily change it later on.
+      const styleTag = document.createElement('style');
+      styleTag.id = THEME_STYLE_ID;
+      styleTag.innerHTML = theme === 'light' ? HLJSLightTheme : HLJSDarkTheme;
+      document.head.appendChild(styleTag);
+    }
+  };
+
+  React.useEffect(() => {
+    handleThemeChange(colorMode);
+  }, [colorMode]);
 
   /**
    * If the language prop is provided, we'll try to override the language
@@ -79,11 +110,11 @@ export const HighlightedMarkdown = (props: HighlightedMarkdownProps) => {
       dangerouslySetInnerHTML={{
         __html: sanitizedHtml,
       }}
-      sx={{
+      sx={(theme) => ({
         '& .hljs': {
           color: theme.color.offBlack,
         },
-      }}
+      })}
       className={`formatted-text ${className}`}
       ref={rootRef}
     />
